requirements.txt
streamlit
pandas
numpy
matplotlib
boto3
seaborn
python-dotenv
overpy
osmnx
folium
streamlit_folium
plotly